[react-restaurant] Follow-up place options in hook
